pkg/util/flag: move to k8s.io/apiserver
