+++ federation/cmd/federation-apiserver/apiserver.go
@@ -24,9 +24,9 @@ import (
 	"os"
 	"time"
 
+	"k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/kubernetes/federation/cmd/federation-apiserver/app"
 	"k8s.io/kubernetes/federation/cmd/federation-apiserver/app/options"
-	"k8s.io/kubernetes/pkg/util/flag"
 	"k8s.io/kubernetes/pkg/util/logs"
 	"k8s.io/kubernetes/pkg/version/verflag"
