+++ federation/cmd/federation-controller-manager/app/options/options.go
@@ -25,10 +25,10 @@ import (
 	"github.com/spf13/pflag"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	utilflag "k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/kubernetes/federation/pkg/dnsprovider"
 	"k8s.io/kubernetes/pkg/apis/componentconfig"
 	"k8s.io/kubernetes/pkg/client/leaderelection"
-	utilflag "k8s.io/kubernetes/pkg/util/flag"
 )
 
 type ControllerManagerConfiguration struct {
