+++ pkg/kubectl/cmd/config/create_cluster.go
@@ -25,11 +25,11 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/client-go/tools/clientcmd"
 	clientcmdapi "k8s.io/client-go/tools/clientcmd/api"
 	"k8s.io/kubernetes/pkg/kubectl/cmd/templates"
 	cmdutil "k8s.io/kubernetes/pkg/kubectl/cmd/util"
-	"k8s.io/kubernetes/pkg/util/flag"
 )
 
 type createClusterOptions struct {
